Add floor to unit characteristics

+++ The-House-Game/Assets/Scripts/UI/UIController.cs
@@ -31,9 +31,9 @@ public class UIController : MonoBehaviour
 
     public void ShowUnitInfo(Unit unit)
     {
-        infoField.transform.Find("UnitField/Health/Value").GetComponent<TMPro.TextMeshProUGUI>().text = unit.GetHealth().ToString();
-        infoField.transform.Find("UnitField/Strength/Value").GetComponent<TMPro.TextMeshProUGUI>().text = unit.CalculateTrueDamage().ToString();
-        infoField.transform.Find("UnitField/Speed/Value").GetComponent<TMPro.TextMeshProUGUI>().text = unit.getSpeed().ToString();
+        infoField.transform.Find("UnitField/Health/Value").GetComponent<TMPro.TextMeshProUGUI>().text = ((int)unit.GetHealth()).ToString();
+        infoField.transform.Find("UnitField/Strength/Value").GetComponent<TMPro.TextMeshProUGUI>().text = ((int)unit.CalculateTrueDamage()).ToString();
+        infoField.transform.Find("UnitField/Speed/Value").GetComponent<TMPro.TextMeshProUGUI>().text = ((int)unit.getSpeed()).ToString();
         infoField.transform.Find("UnitField").gameObject.SetActive(true);
         int i = 0;
         foreach (float hp in unit.GetAllHealths())
@@ -41,7 +41,7 @@ public class UIController : MonoBehaviour
             var singleUnit = Instantiate(unitInfoPrefab, groupInfo.transform);
             var newPos = new Vector3(singleUnit.transform.localPosition.x + i * 100, singleUnit.transform.localPosition.y, singleUnit.transform.localPosition.z);
             singleUnit.transform.SetLocalPositionAndRotation(newPos, singleUnit.transform.rotation);
-            singleUnit.transform.Find("Health").GetComponent<TMPro.TextMeshProUGUI>().text = hp.ToString();
+            singleUnit.transform.Find("Health").GetComponent<TMPro.TextMeshProUGUI>().text = ((int)hp).ToString();
             ++i;
         }
     }
